Refactor out into a new file.
Also refactor loading to a method.
